Moved ArticleProcessor to module
