+++ AndroidApp/.idea/gradle.xml
@@ -11,6 +11,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/ArticleProcessor" />
             <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
