+++ AndroidApp/app/build.gradle
@@ -56,4 +56,5 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
     implementation files('libs/RSSParser.jar')
+    implementation project(path: ':ArticleProcessor')
 }
