+++ AndroidApp/app/src/main/java/com/niilopoutanen/rss_feed/activities/ArticleActivity.java
@@ -35,7 +35,7 @@ import com.niilopoutanen.rss_feed.R;
 import com.niilopoutanen.rss_feed.adapters.ArticleAdapter;
 import com.niilopoutanen.rss_feed.models.ArticleQuoteSpan;
 import com.niilopoutanen.rss_feed.models.Preferences;
-import com.niilopoutanen.rss_feed.utils.ArticleProcessor;
+import com.niilopoutanen.articleprocessor.*;
 import com.niilopoutanen.rss_feed.utils.PreferencesManager;
 
 import net.dankito.readability4j.Article;
