+++ AndroidApp/settings.gradle
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "RSS-Feed"
 include ':app'
+include ':ArticleProcessor'
